requirements.txt
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
google-generativeai==0.3.2